+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Add item.effectiveLocationPrimaryServicePointName as staff slip token in Settings. Ref CIRC-1785
 * Leverage cookie-based authentication in all API requests. Refs UICIRC-881.
 * Also support `circulation` `14.0`. Refs UICIRC-935.
+* Add request.requestDate as staff slip token in Settings. Ref CIRC-1784
 
 ## [8.0.1](https://github.com/folio-org/ui-circulation/tree/v8.0.1) (2023-03-07)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v8.0.0...v8.0.1)